poisto onnistuu vain kirjautuneella oikealla käyttäjällä: versio 4.21

+++ osa4/blogilista-backend/controllers/blogs.js
@@ -53,8 +53,24 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+
+  if ( blog.user.toString() === user._id.toString() ) {
+    await Blog.findByIdAndRemove(request.params.id)
+    return response.status(204).end()
+  }
+
+  return response.status(400).json({
+    error: 'user is not the one who added the blog: blog can not be removed'
+  })
 })
 
 blogsRouter.put('/:id', async (request, response) => {
